ref: Delete old input module
