+++ internal/display/input/postingaccount_test.go
@@ -10,8 +10,8 @@ import (
 	. "github.com/vitorqb/addledger/internal/display/input"
 	"github.com/vitorqb/addledger/internal/display/widgets"
 	eventbusmod "github.com/vitorqb/addledger/internal/eventbus"
-	"github.com/vitorqb/addledger/internal/input"
 	"github.com/vitorqb/addledger/internal/listaction"
+	"github.com/vitorqb/addledger/internal/userinput"
 	. "github.com/vitorqb/addledger/mocks/controller"
 )
 
@@ -45,7 +45,7 @@ func TestNewPostingAccountField(t *testing.T) {
 			name: "Sends OnPostingAccountSelectedFromContext msg",
 			run: func(t *testing.T, c *testcontext) {
 				c.controller.EXPECT().OnPostingAccountChanged("FOO")
-				c.controller.EXPECT().OnPostingAccountDone(input.Context)
+				c.controller.EXPECT().OnPostingAccountDone(userinput.Context)
 				c.postingAccount.SetText("FOO")
 				event := tcell.NewEventKey(tcell.KeyEnter, ' ', tcell.ModNone)
 				c.postingAccount.InputHandler()(event, func(tview.Primitive) {})
@@ -78,7 +78,7 @@ func TestNewPostingAccountField(t *testing.T) {
 			name: "Calls OnPostingAccountDone on Ctrl+J",
 			run: func(t *testing.T, c *testcontext) {
 				c.controller.EXPECT().OnPostingAccountChanged("FOO")
-				c.controller.EXPECT().OnPostingAccountDone(input.Input)
+				c.controller.EXPECT().OnPostingAccountDone(userinput.Input)
 				c.postingAccount.SetText("FOO")
 				event := tcell.NewEventKey(tcell.KeyCtrlJ, ' ', tcell.ModNone)
 				c.postingAccount.InputHandler()(event, func(tview.Primitive) {})
